ref(expo-plugin): Split utils to logger, version and utils
